[KAN-213] feat(payment): FCM 초기 세팅 완료
