+++ client/package.json
@@ -20,6 +20,7 @@
     "@fullcalendar/timegrid": "^6.1.15",
     "@headlessui/react": "^2.1.2",
     "axios": "^1.7.2",
+    "firebase": "^10.12.3",
     "path-browserify": "^1.0.1",
     "qrcode.react": "^3.1.0",
     "react": "^18.3.1",
