+++ client/src/routes/root.jsx
@@ -5,6 +5,7 @@ import Login from '@/pages/LoginPage'; // 로그인
 import Signup from '@/pages/SignupPage'; // 회원가입
 import VendorIndex from '@/pages/vendor/VenIndexPage'; // 고객 중첩 라우팅
 import MemberIndex from '@/pages/member/MemIndexPage'; // 회원 중첩 라우팅
+import Test from '@/labs/Test';
 
 const root = createBrowserRouter([
   {
@@ -28,6 +29,10 @@ const root = createBrowserRouter([
     path: 'member',
     element: <MemberIndex />,
     children: memberRoute(),
+  },
+  {
+    path: 'test',
+    element: <Test />,
   },
 ]);
