--- server-message/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'cmsplusmessage'
