--- server-message/src/main/java/kr/or/kosa/cmsplusmessage/KafkaTestDto.java
@@ -1,19 +1,0 @@
-//package kr.or.kosa.cmsplusmessage;
-//
-//import lombok.Getter;
-//import lombok.Setter;
-//import lombok.ToString;
-//
-//import java.io.Serializable;
-//
-//
-//@Getter
-//@Setter
-//@ToString
-//public class KafkaTestDto implements Serializable {
-//
-//    private String phone;
-//    private String message;
-//
-//}
-//
