--- server-message/src/main/java/kr/or/kosa/cmsplusmessage/dto/EmailMessageDto.java
@@ -1,23 +1,0 @@
-package kr.or.kosa.cmsplusmessage.dto;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import kr.or.kosa.cmsplusmessage.MessageSendMethod;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class EmailMessageDto extends MessageDto {
-
-    private String emailAddress;
-
-    @JsonCreator
-    public EmailMessageDto(@JsonProperty("text") String text, @JsonProperty("emailAddress") String emailAddress) {
-        super(MessageSendMethod.EMAIL, text);
-        this.emailAddress = emailAddress;
-    }
-
-}
