--- server-message/src/test/java/kr/or/kosa/cmsplusmessage/CmsplusmessageApplicationTests.java
@@ -1,13 +1,0 @@
-package kr.or.kosa.cmsplusmessage;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CmsplusmessageApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
