+++ server-messaging/src/main/java/kr/or/kosa/cmsplusmessaging/config/EmailConfig.java
@@ -1,0 +1,38 @@
+package kr.or.kosa.cmsplusmessaging.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.JavaMailSenderImpl;
+
+import java.util.Properties;
+
+@Configuration
+public class EmailConfig {
+
+    @Value("${email.emailAddress}")
+    private String emailAddress;
+    @Value("${email.password}")
+    private String password;
+    @Value("${email.username}")
+    private String username;
+
+    @Bean
+    public JavaMailSender getJavaMailSender() {
+        JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
+        mailSender.setHost("smtp.gmail.com");
+        mailSender.setPort(587);
+        mailSender.setUsername(username);
+        mailSender.setPassword(password);
+
+        Properties props = mailSender.getJavaMailProperties();
+        props.put("mail.transport.protocol", "smtp");
+        props.put("mail.smtp.auth", "true");
+        props.put("mail.smtp.starttls.enable", "true");
+        props.put("mail.smtp.timeout", "5000");
+        return mailSender;
+
+    }
+
+}
