+++ server-messaging/src/test/java/kr/or/kosa/cmsplusmessaging/CmsplusmessagingApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.or.kosa.cmsplusmessaging;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CmsplusmessagingApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
