+++ server-payment/src/main/java/kr/or/kosa/cmspluspayment/CmsplusPaymentApplication.java
@@ -1,0 +1,13 @@
+package kr.or.kosa.cmspluspayment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CmsplusPaymentApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CmsplusPaymentApplication.class, args);
+	}
+
+}
