+++ server-payment/src/main/java/kr/or/kosa/cmspluspayment/dto/PaymentDto.java
@@ -1,0 +1,14 @@
+package kr.or.kosa.cmspluspayment.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class PaymentDto {
+
+    private String message;
+
+}
