+++ server-payment/src/test/java/kr/or/kosa/cmspluspayment/CmsplusPaymentApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.or.kosa.cmspluspayment;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CmsplusPaymentApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
