+++ server/src/main/java/kr/or/kosa/cmsplusmain/domain/messaging/dto/TestDto.java
@@ -1,0 +1,14 @@
+package kr.or.kosa.cmsplusmain.domain.messaging.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class TestDto {
+
+    private String message;
+
+}
